server/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bitarray==3.8.0
bitstring==4.3.1
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
click-default-group==1.2.4
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.8.0
esptool==5.1.0
exceptiongroup==1.3.0
fastapi==0.119.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
intelhex==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
mpremote==1.26.1
multidict==6.7.0
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
postgrest==2.24.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
pyserial==3.5
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.24.0
reedsolo==1.7.0
rich==14.2.0
rich-click==1.9.4
six==1.17.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-utils==3.39
starlette==0.48.0
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
tabulate==0.9.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.22.0
